manual_deploy_temp/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
python-dotenv
crewai
chromadb
pysqlite3-binary
openai
boto3
requests
tiktoken
numpy
PyYAML
